requirements.txt
pytest
allure-pytest
selene>=2.0.0rc3.post1
